+++ ChangeLog.md
@@ -15,6 +15,8 @@ All notable changes to this project will be documented in this file.
 - :openai module support for assistant run api (...)
 - :openai module support for assistant run steps api (...)
 
+- IPC: implemented optional payload message compression
+
 ### Removed
 
 - gitpod support